+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/oauth2/TokenGenerationService.kt
@@ -19,7 +19,7 @@ import at.asitplus.wallet.lib.oidvci.DefaultNonceService
 import at.asitplus.wallet.lib.oidvci.MapStore
 import at.asitplus.wallet.lib.oidvci.NonceService
 import at.asitplus.wallet.lib.oidvci.OAuth2Exception.InvalidDpopProof
-import io.github.aakira.napier.Napier
+import at.asitplus.wallet.lib.oidvci.OAuth2Exception.UseDpopNonce
 import kotlin.time.Clock
 import kotlin.time.Clock.System
 import kotlin.time.Duration.Companion.days
@@ -71,7 +71,7 @@ class JwtTokenGenerationService(
         scope: String?,
         validatedClientKey: JsonWebKey?,
     ): TokenResponseParameters = if (httpRequest?.dpop == null) {
-        throw InvalidDpopProof("no DPoP header value")
+        throw InvalidDpopProof("no dpop proof in header")
     } else {
         val notBefore = clock.now().truncateToSeconds()
         TokenResponseParameters(